+++ src/samotech_iptv/domain/entities/__init__.py
@@ -9,6 +9,8 @@ Or directly from the sub-module::
     from samotech_iptv.domain.entities.channel import Channel
 """
 
+from samotech_iptv.domain.entities.account_info import AccountInfo
+from samotech_iptv.domain.entities.catchup_event import CatchupEvent
 from samotech_iptv.domain.entities.category import Category
 from samotech_iptv.domain.entities.channel import Channel
 from samotech_iptv.domain.entities.epg_entry import EPGEntry
@@ -18,12 +20,16 @@ from samotech_iptv.domain.entities.history import History
 from samotech_iptv.domain.entities.movie import Movie
 from samotech_iptv.domain.entities.playlist import Playlist
 from samotech_iptv.domain.entities.provider import Provider
+from samotech_iptv.domain.entities.provider_session import ProviderSession
 from samotech_iptv.domain.entities.season import Season
 from samotech_iptv.domain.entities.series import Series
+from samotech_iptv.domain.entities.server_info import ServerInfo
 from samotech_iptv.domain.entities.stream import Stream
 from samotech_iptv.domain.entities.xmltv_binding import XMLTVBinding, XMLTVChannelMapping
 
 __all__ = [
+    "AccountInfo",
+    "CatchupEvent",
     "Channel",
     "Category",
     "Playlist",
@@ -33,6 +39,8 @@ __all__ = [
     "Episode",
     "Stream",
     "Provider",
+    "ProviderSession",
+    "ServerInfo",
     "EPGEntry",
     "Favorite",
     "History",